// backend/apps/parcel-service/src/parcel/enums/parcel.enums.ts
export enum ParcelStatus {
  BOOKED = 'BOOKED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  CONFIRMED = 'CONFIRMED',
  DISPATCHED = 'DISPATCHED',
  IN_TRANSIT = 'IN_TRANSIT',
  ARRIVED = 'ARRIVED',
  READY_FOR_PICKUP = 'READY_FOR_PICKUP',
  DELIVERED = 'DELIVERED',
  RETURN_TO_SENDER = 'RETURN_TO_SENDER',
}